test(frontend): move the sidebar tests into @agenta/navigation
They tested pure functions that live in the package — grouping, the date
buckets, the recency sort — through web/oss's thin re-export, which is the
wrong home for them and put 190 lines of test into an app the feature does not
touch. The package gains the vitest setup @agenta/entities already uses;
web/oss returns to its original three.

Also drops the navigation from an agent heading and its plumbing (the group's
`path`, the `link` it resolved to, and `groupedChildren`'s now-unused
projectURL). The caret loses its hover fill too — the row already answers with
a colour change, and a pill behind the caret made a heading look like a button.

Pinned is collapsible again. It was pinned into the expanded set on every read,
so the caret could never remove it; its head start is now the stored record's
default, which the first toggle overwrites. The seed that opens an active
session's group moved to module scope for the same reason: the rail and the
drawer each ran the hook, so one instance kept re-opening what the user had
just closed through the other.

And pinning a session no longer blanks the rail: a new pinned-query key reports
its first flight as pending, which emptied the list for a frame before it came
back reordered. Pending now means pending with nothing to show.
