+++ web/packages/agenta-navigation/.gitignore
@@ -1,0 +1,3 @@
+# Generated by Vitest — do not commit
+test-results/
+coverage/
