+++ web/packages/agenta-navigation/package.json
@@ -9,7 +9,11 @@
         "build": "tsc --noEmit",
         "types:check": "tsc --noEmit",
         "lint": "eslint --config ./eslint.config.mjs src/",
-        "check": "pnpm run types:check && pnpm run lint"
+        "check": "pnpm run types:check && pnpm run lint",
+        "test": "pnpm run test:unit",
+        "test:unit": "vitest run",
+        "test:watch": "vitest",
+        "test:coverage": "vitest run --coverage"
     },
     "exports": {
         ".": "./src/index.ts"
@@ -31,9 +35,11 @@
         "@tanstack/react-query": "5.100.9",
         "@types/node": "^20.19.20",
         "@types/react": "^19.0.10",
+        "@vitest/coverage-v8": "^4.1.4",
         "jotai": "^2.15.0",
         "jotai-family": "^1.0.1",
         "react": "^19.0.0",
-        "typescript": "^5.9.3"
+        "typescript": "^5.9.3",
+        "vitest": "^4.1.4"
     }
 }
